requirements.txt
httpx
openpyxl
pandas
pandas-gbq
python-amazon-sp-api
python-dotenv
